requirements.txt
streamlit
moviepy
librosa
mediapipe
google-generativeai
numpy
opencv-python
python-dotenv
